Rewrite Hotswap2 tests to MUnit
